refactor: 멘티 이벤트 참여 신청 도메인 분리
