+++ src/main/java/org/devcourse/resumeme/business/event/controller/MenteeToEventController.java
@@ -3,7 +3,6 @@ package org.devcourse.resumeme.business.event.controller;
 import lombok.RequiredArgsConstructor;
 import org.devcourse.resumeme.business.event.controller.dto.ApplyToEventRequest;
 import org.devcourse.resumeme.business.event.controller.dto.v2.ApplyUpdateRequest;
-import org.devcourse.resumeme.business.event.service.EventService;
 import org.devcourse.resumeme.business.event.service.MenteeToEventService;
 import org.devcourse.resumeme.business.event.service.vo.AcceptMenteeToEvent;
 import org.devcourse.resumeme.business.event.service.vo.ApplyUpdateVo;
@@ -11,7 +10,12 @@ import org.devcourse.resumeme.business.resume.service.ComponentService;
 import org.devcourse.resumeme.business.resume.service.ResumeService;
 import org.devcourse.resumeme.global.auth.model.jwt.JwtUser;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.PatchMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequiredArgsConstructor
@@ -24,13 +28,11 @@ public class MenteeToEventController {
 
     private final MenteeToEventService applyService;
 
-    private final EventService eventService;
-
     @PostMapping("/{eventId}")
     public void applyEvent(@PathVariable Long eventId, @RequestBody ApplyToEventRequest request, @AuthenticationPrincipal JwtUser user) {
         Long copyResumeId = resumeService.copyResume(request.resumeId());
         componentService.copy(request.resumeId(), copyResumeId);
-        eventService.acceptMentee(new AcceptMenteeToEvent(eventId, user.id(), copyResumeId));
+        applyService.acceptMentee(new AcceptMenteeToEvent(eventId, user.id(), copyResumeId));
     }
 
     @PatchMapping("/{eventId}")
