--- src/main/java/org/devcourse/resumeme/business/event/service/ApplyProvider.java
@@ -1,7 +1,0 @@
-package org.devcourse.resumeme.business.event.service;
-
-public interface ApplyProvider {
-
-    Long getApplyEventCount(Long menteeId);
-
-}
