+++ src/main/java/org/devcourse/resumeme/business/event/service/MenteeToEventService.java
@@ -7,7 +7,9 @@ import org.devcourse.resumeme.business.event.domain.model.ApplimentUpdate;
 import org.devcourse.resumeme.business.event.exception.EventException;
 import org.devcourse.resumeme.business.event.repository.EventRepository;
 import org.devcourse.resumeme.business.event.repository.MenteeToEventRepository;
+import org.devcourse.resumeme.business.event.service.vo.AcceptMenteeToEvent;
 import org.devcourse.resumeme.business.event.service.vo.ApplyUpdateVo;
+import org.devcourse.resumeme.global.exception.ExceptionCode;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -18,7 +20,7 @@ import static org.devcourse.resumeme.global.exception.ExceptionCode.EVENT_NOT_FO
 @Service
 @Transactional
 @RequiredArgsConstructor
-public class MenteeToEventService implements ApplyProvider {
+public class MenteeToEventService {
 
     private final MenteeToEventRepository menteeToEventRepository;
 
@@ -29,19 +31,33 @@ public class MenteeToEventService implements ApplyProvider {
         return menteeToEventRepository.findAllByResumeId(resumeId);
     }
 
-    @Override
-    public Long getApplyEventCount(Long menteeId) {
-        return menteeToEventRepository.findByMenteeId(menteeId).stream()
-                .filter(MenteeToEvent::isAttending)
-                .count();
-    }
-
     public void update(Long eventId, ApplyUpdateVo applyUpdateVo) {
         Event event = eventRepository.findWithApplicantsById(eventId)
                 .orElseThrow(() -> new EventException(EVENT_NOT_FOUND));
 
         ApplimentUpdate model = applyUpdateVo.toModel();
         model.update(event);
+    }
+
+    public void acceptMentee(AcceptMenteeToEvent ids) {
+        checkCanApply(ids.menteeId());
+        Event event = eventRepository.findWithLockById(ids.eventId())
+                .orElseThrow();
+
+        event.acceptMentee(ids.menteeId(), ids.resumeId());
+    }
+
+    private void checkCanApply(Long menteeId) {
+        Long menteeApplyCount = getApplyEventCount(menteeId);
+        if (menteeApplyCount > 0) {
+            throw new EventException(ExceptionCode.DUPLICATE_APPLICATION_EVENT);
+        }
+    }
+
+    private Long getApplyEventCount(Long menteeId) {
+        return menteeToEventRepository.findByMenteeId(menteeId).stream()
+                .filter(MenteeToEvent::isAttending)
+                .count();
     }
 
 }
